// src/app/components/Models/UserDetails.ts
export class UserDetails {
    Designation?: string;
    EmailId?: string;
    IsManager?: boolean;
    LastUpdateBy?: string;
    LastUpdatedOn?: string;
    ManagerCode?: string;
    ManagerEmailId?: string;
    ManagerId?: number;
    Name?: string;
    RoleCode?: string;
    Status?: number;
    Surname?: string;
    UserCode?: string;
    UserId: number;
    DOB?: string;
    FullNameWithDesignation?: string;
}